Hot fix

+++ 0x00-hello_world/6-size.c
@@ -10,15 +10,14 @@ int main(void)
 {
 	int g;
 	long int h;
-	long int i;
-
+	long long int i;
 	float j;
 	char k;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(k));
-	printf("Size of an int: %lu byte(s)\n", sizeof(g));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(h));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(i));
-	printf("Size of a float: %lu byte(s)\n", sizeof(j));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(k));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(g));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(h));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(j));
 	return (0);
 }
